+++ src/nim/function_export/function_export.h
@@ -3,7 +3,7 @@
 #define DLLAPI __declspec( dllexport )
 
 // The following lines the define the different capi function calls exposed to the dll
-DLLAPI float __cdecl cadd(float a, float b);
+DLLAPI double __cdecl cadd(double a, double b);
 DLLAPI int __cdecl get_name_size();
 DLLAPI void __cdecl get_name(char* buffer);
